Device works as both CDC + HID
This is in a working state, device will transition between CDC and HID with the holding down of a button. Once the device is in CDC, it is fully programmable between 4 different logins. The device acts as a keyboard once it is put into HID mode. Selecting between logins is done by pushing the same button used to switch between CDC and HID. Pushing the other button will send the Auto Login key presses.